mcp-server/pyproject.toml
[project]
name = "mcp-excel-server"
version = "1.0.0"
description = "MCP Excel Service for Excel file manipulation via Microsoft Graph API"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.40.0",
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
include = ["server.py"]
